refactor: update Calculator and MainFrame from int to long

+++ src/main/java/com/bruno/calculadora/Calculator.java
@@ -4,41 +4,41 @@ public class Calculator {
 
     // ESCENARIO 1: El cliente ingresa cuántos Robux LIMPIOS quiere (Llegan)
 
-    public int calculateGamepassFromClean(int cleanRobux) {
+    public long calculateGamepassFromClean(long cleanRobux) {
         double gamepassRobux = cleanRobux / 0.7;
-        return (int) Math.ceil(gamepassRobux);
+        return (long) Math.ceil(gamepassRobux);
 
     }
 
-    public int calculatePriceFromClean(int cleanRobux, Seller seller) {
-        int gamepassPrice = calculateGamepassFromClean(cleanRobux);
+    public long  calculatePriceFromClean(long cleanRobux, Seller seller) {
+        long gamepassPrice = calculateGamepassFromClean(cleanRobux);
         double sellerPrice = seller.getPricePer80Robux();
         double exactPrice = ( gamepassPrice * sellerPrice ) / 80.0;
-        return (int) Math.round(exactPrice);
+        return (long) Math.round(exactPrice);
     }
 
     // ESCENARIO 2: El cliente ingresa de cuánto es el PASE
 
-    public int calculateCleanFromGamepass(int gamepassRobux) {
-        return (int) (gamepassRobux * 0.7);
+    public long calculateCleanFromGamepass(long gamepassRobux) {
+        return (long) (gamepassRobux * 0.7);
     }
 
-    public int calculatePriceFromGamepass(int gamepassRobux, Seller seller) {
+    public long calculatePriceFromGamepass(long gamepassRobux, Seller seller) {
         double sellerPrice = seller.getPricePer80Robux();
         double exactPrice = ( gamepassRobux * sellerPrice) / 80.0;
-        return (int) Math.round(exactPrice);
+        return (long) Math.round(exactPrice);
     }
 
     // ESCENARIO 3: El cliente ingresa cuánta PLATA (Pesos) tiene
 
-    public int calculateGamepassFromPrice(double priceInPesos, Seller seller) {
+    public long calculateGamepassFromPrice(double priceInPesos, Seller seller) {
         double sellerPrice = seller.getPricePer80Robux();
         double exactPrice = ( priceInPesos * 80.0 ) / sellerPrice;
-        return (int) Math.floor(exactPrice);
+        return (long) Math.floor(exactPrice);
     }
 
-    public int calculateCleanFromPrice(double priceInPesos, Seller seller) {
-        int gamepassAmount = calculateGamepassFromPrice(priceInPesos, seller);
+    public long calculateCleanFromPrice(double priceInPesos, Seller seller) {
+        long gamepassAmount = calculateGamepassFromPrice(priceInPesos, seller);
         return calculateCleanFromGamepass(gamepassAmount);
     }
